Unit Animations and Warning Fixing

+++ Order_227/Motor2D/App.cpp
@@ -88,7 +88,7 @@ bool App::Awake()
 	{
 		// self-config
 		ret = true;
-		float fps = frame_time.Read();
+		float fps = (float)frame_time.Read();
 		app_config = config.child("app");
 		title.assign(app_config.child("title").child_value());
 		organization.assign(app_config.child("organization").child_value());
